refactored TopCoins.js to CoinsCard.js as component and handover the properties
